Automatic formatting of Social media release announcement
A python script that depends on packages (in requirements.txt)
* requests
* markdown

able to format a social media post announcing our releases.
The script will pick, for each operator, the latest tag name and related
URL to release note, and format a message in raw markdown and HTML.

suggested use

$ python -m venv venv
$ source venv/bin/activate
$ pip install -r requirements.txt
$ python ./product_release_post.py

Limitations
There's a (currently unhandled) rate limit to the use of APIs. It is
enough for around 10 retries. When reached the script fails with error

```
KeyError: 'tag_name'
```

Signed-off-by: Carlo Lobrano <c.lobrano@gmail.com>